# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/admin/confirmed_booking/{booking_id}": {
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Admin Confirmed Booking",
                "operationId": "adminConfirmedBooking",
                "parameters": [
                    {
                        "type": "string",
                        "description": "BookingID",
                        "name": "booking_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.listBookingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/admin/list_booking_confirmed/": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Admin Get List Booking Confirmed",
                "operationId": "adminGetListBookingConfirmed",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.listBookingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/admin/list_booking_validated/": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Admin Get List Booking Validated",
                "operationId": "adminGetListBookingValidated",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.listBookingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/admin/users/{username}/bookings/{homestay_booking}/{booking_id}/checkout": {
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Admin CheckOut Booking",
                "operationId": "checkoutBooking",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserBooking",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "HomestayBooking",
                        "name": "homestay_booking",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "BookingID",
                        "name": "booking_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/users/{username}/bookings/{homestay_booking}": {
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User Create Booking",
                "operationId": "createBooking",
                "parameters": [
                    {
                        "description": "createBookingRequest data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createBookingRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "UserBooking",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "HomestayBooking",
                        "name": "homestay_booking",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.BookingTxResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/users/{username}/bookings/{homestay_booking}/{booking_id}/cancel": {
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User Cancel Booking",
                "operationId": "cancelBooking",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserBooking",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "HomestayBooking",
                        "name": "homestay_booking",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "BookingID",
                        "name": "booking_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.createBookingRequest": {
            "type": "object",
            "required": [
                "checkin_date",
                "number_of_day",
                "number_of_guest",
                "promotion_id"
            ],
            "properties": {
                "checkin_date": {
                    "type": "string"
                },
                "number_of_day": {
                    "type": "integer"
                },
                "number_of_guest": {
                    "type": "integer",
                    "minimum": 1
                },
                "promotion_id": {
                    "type": "string"
                }
            }
        },
...
        "api.listBookingResponse": {
            "type": "object",
            "properties": {
                "list_booking": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "booking": {
                                "$ref": "#/definitions/db.Booking"
                            },
                            "detail_payment": {
                                "$ref": "#/definitions/db.DetailPayment"
                            },
                            "homestay_booking": {
                                "$ref": "#/definitions/db.Homestay"
                            },
                            "payment": {
                                "$ref": "#/definitions/db.Payment"
                            },
                            "user_booking": {
                                "$ref": "#/definitions/api.userResponse"
                            }
                        }
                    }
                }
            }
        },
...
        "db.Booking": {
            "type": "object",
            "properties": {
                "booking_date": {
                    "type": "string"
                },
                "booking_id": {
                    "type": "string"
                },
                "checkin_date": {
                    "type": "string"
                },
                "checkout_date": {
                    "type": "string"
                },
                "homestay_booking": {
                    "type": "integer"
                },
                "number_of_guest": {
                    "type": "integer"
                },
                "promotion_id": {
                    "type": "string"
                },
                "service_fee": {
                    "description": "must be positive",
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "tax": {
                    "description": "must be positive",
                    "type": "number"
                },
                "user_booking": {
                    "type": "string"
                }
            }
        },
        "db.BookingTxResult": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/db.Booking"
                },
                "detail_payment": {
                    "$ref": "#/definitions/db.DetailPayment"
                },
                "homestay_booking": {
                    "$ref": "#/definitions/db.Homestay"
                },
                "payment": {
                    "$ref": "#/definitions/db.Payment"
                },
                "user_booking": {
                    "$ref": "#/definitions/db.userResponse"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}